[Change] 一部ボタンのUIを変更

+++ app/src/components/modals/ShowEventModal.jsx
@@ -1,31 +1,34 @@
 import { useRef } from 'react'
 import { ModalWindow } from '/src/components/modals/ModalWindow'
-import { PencilSquareIcon, XCircleIcon } from '@heroicons/react/24/outline'
+import { PencilSquareIcon, XCircleIcon, TrashIcon } from '@heroicons/react/24/outline'
 import { destroyCalendarEvent } from '/src/lib/api/calendarEvent'
 
 export const ShowEventModal = ({ modal, icon, calendarEvent, editEventModal, updateCalendar }) => {
   const isProc = useRef(false); //送信処理の管理
 
   const eventDestroy = async () => {
-    if (isProc.current) return;
-    isProc.current = true;
+    if (window.confirm('この予定を削除しますします。よろしいですか？')) {
+      if (isProc.current) return;
+      isProc.current = true;
 
-    await destroyCalendarEvent({
-      id: calendarEvent.id
-    }).then(() => {
-      updateCalendar().then(() => {
-        modal.close();
-        calendarEvent.init();
+      await destroyCalendarEvent({
+        id: calendarEvent.id
+      })
+      .then(() => {
+        updateCalendar().then(() => {
+          modal.close();
+          calendarEvent.init();
+        });
+      })
+      .catch((err) => {
+        // エラーメッセージのアラートを表示
+        const errorMessages = err.response.data.errors;
+        alert(errorMessages.join('\n'));
+      })
+      .finally(() => {
+        isProc.current = false;
       });
-    })
-    .catch((err) => {
-      // エラーメッセージのアラートを表示
-      const errorMessages = err.response.data.errors;
-      alert(errorMessages.join('\n'));
-    })
-    .finally(() => {
-      isProc.current = false;
-    });
+    }
   };
 
   return (
@@ -66,13 +69,11 @@ export const ShowEventModal = ({ modal, icon, calendarEvent, editEventModal, upd
         </div>
       </div>
       <div className='flex justify-between mt-10'>
-        <button type='button' className='button button-danger'
-          onClick={() => {
-            if (window.confirm('この予定を削除しますします。よろしいですか？')) eventDestroy();
-          }}
+        <button type='button' className='button button-danger button-rounded'
+          onClick={eventDestroy}
         >
-          <XCircleIcon className='h-6 w-6' />
-          予定を削除
+          <TrashIcon className='h-6 w-6' />
+          
         </button>
 
         <button type='button' className='button'
